+++ Moco/Moco/MocoGoal/MocoPeriodicityGoal.h
@@ -22,63 +22,71 @@
 
 namespace OpenSim {
 
-/// Create pair of elements for use with a MocoPeriodicityGoal.
+/// Create pair of variables for use with a MocoPeriodicityGoal.
 class OSIMMOCO_API MocoPeriodicityGoalPair : public Object {
     OpenSim_DECLARE_CONCRETE_OBJECT(MocoPeriodicityGoalPair, Object);
 
 public:
     OpenSim_DECLARE_PROPERTY(
-            first, std::string, "First element of the pair.");
+            initial, std::string, "Initial variable of the pair.");
     OpenSim_DECLARE_PROPERTY(
-            second, std::string, "Second element of the pair.");
+            final, std::string, "Final variable of the pair.");
 
     MocoPeriodicityGoalPair();
-    MocoPeriodicityGoalPair(std::string name);
+    MocoPeriodicityGoalPair(std::string initial, std::string final);
+    MocoPeriodicityGoalPair(std::string initialIsFinal);
 
 private:
     void constructProperties();
 
 };
 
-/// This goal enforces equality between initial and final state values in the
-/// optimal control problem. The initial and final states can belong to
+/// This goal enforces equality between initial and final variable values in
+/// the optimal control problem. The initial and final values can belong to
 /// separate continuous variables or the same continuous variable, as long as
-/// they are the same type of variable. Value pairs are specified via a
-/// MocoPeriodicityGoalPair, where the first element of the pair indicates the
-/// initial state variable and the second element indicates the final state
-/// variable. Set the first and second elements to the same variable to enforce
-/// periodicity on a single continuous variable. Only state and control
-/// continuous variable pairs are supported, and are specified via the
-/// 'state_pairs' and 'control_pairs' properties.
+/// they are the same type of variable (e.g., state or control). Value pairs
+/// are specified via a MocoPeriodicityGoalPair, where the initial variable of
+/// the pair indicates the initial state/control variable and the final
+/// variable indicates the final state/control variable. Set the initial and
+/// final variables to the same variable to enforce periodicity on a single
+/// continuous variable. Only state and control continuous variable pairs are
+/// supported, and are specified via the 'state_pairs' and 'control_pairs'
+/// properties.
 ///
-/// Imposing periodicity of pelvis tilt values, hip speeds, and hamstrings
-/// controls after half a gait cycle assuming symmetry of the gait cycle (i.e.,
-/// right and left leg values, and inversely, should be the same at the
-/// beginning and half of the gait cycle, respectively):
+/// To impose bilateral symmetry in a walking simulation, we can simulate over
+/// half a gait cycle and impose periodic constraints. For bilateral variables
+/// (e.g., hip flexion speeds and hamstrings controls), the constraints should
+/// enforce that right and left values, and inversely, should be the same at
+/// the beginning and half of the gait cycle, respectively. For the other
+/// variables (e.g., pelvis tilt values), the constraints should enforce that
+/// the values should be the same at the beginning and half of the gait cycle.
+/// The example code below illustrates how to enforce the aforementioned
+/// constraints cwith different constructors.
 /// @code
-/// auto* periodicGoal = problem.addGoal<MocoPeriodicityGoal>("periodicGoal");
-/// MocoPeriodicityGoalPair pair_pelvis_tilt_values;
-/// pair_pelvis_tilt_values.set_first("/jointset/Pelvis/pelvis_tilt/value");
-/// pair_pelvis_tilt_values.set_second("/jointset/Pelvis/pelvis_tilt/value");
-/// MocoPeriodicityGoalPair pair_hip1_speeds;
-/// pair_hip1_speeds.set_first("/jointset/hip_l/hip_flexion_l/speed");
-/// pair_hip1_speeds.set_second("/jointset/hip_r/hip_flexion_r/speed");
-/// MocoPeriodicityGoalPair pair_hip2_speeds;
-/// pair_hip2_speeds.set_first("/jointset/hip_r/hip_flexion_r/speed");
-/// pair_hip2_speeds.set_second("/jointset/hip_l/hip_flexion_l/speed");
+/// periodicGoal = problem.addGoal<MocoPeriodicityGoal>("periodicGoal");
+/// @endcode
+/// Periodic contraint for the pelvis tilt values:
+/// @code
+/// periodicGoal->addStatePair({ "/jointset/ground_pelvis/pelvis_tilt/value"});
+/// @endcode
+/// Periodic contraints for the hip flexion speeds:
+/// @code
+/// periodicGoal->addStatePair({"/jointset/hip_l/hip_flexion_l/speed",
+///         "/jointset/hip_r/hip_flexion_r/speed"});
+/// periodicGoal->addStatePair({"/jointset/hip_r/hip_flexion_r/speed",
+///         "/jointset/hip_l/hip_flexion_l/speed"});
+/// @endcode
+/// Periodic contraints for the hamstrings controls:
+/// @code
 /// MocoPeriodicityGoalPair pair_hamstrings1;
-/// pair_hamstrings1.set_first("/hamstrings_r");
-/// pair_hamstrings1.set_second("/hamstrings_l");
-/// MocoPeriodicityGoalPair pair_hamstrings2;
-/// pair_hamstrings2.set_first("/hamstrings_l");
-/// pair_hamstrings2.set_second("/hamstrings_r");
-/// periodicGoal->append_state_pairs(pair_pelvis_tilt_values);
-/// periodicGoal->append_state_pairs(pair_hip1_speeds);
-/// periodicGoal->append_state_pairs(pair_hip2_speeds);
+/// pair_hamstrings1.set_initial("/hamstrings_r");
+/// pair_hamstrings1.set_final("/hamstrings_l");
 /// periodicGoal->append_control_pairs(pair_hamstrings1);
+/// MocoPeriodicityGoalPair pair_hamstrings2;
+/// pair_hamstrings2.set_initial("/hamstrings_l");
+/// pair_hamstrings2.set_final("/hamstrings_r");
 /// periodicGoal->append_control_pairs(pair_hamstrings2);
 /// @endcode
-///
 /// This is an endpoint constraint goal by default.
 /// @ingroup mocogoal
 class OSIMMOCO_API MocoPeriodicityGoal : public MocoGoal {
@@ -93,6 +101,13 @@ public:
     MocoPeriodicityGoal();
     MocoPeriodicityGoal(std::string name) : MocoGoal(std::move(name)) {
         constructProperties();
+    }
+
+    void addStatePair(MocoPeriodicityGoalPair pair) {
+        append_state_pairs(std::move(pair));
+    }
+    void addControlPair(MocoPeriodicityGoalPair pair) {
+        append_control_pairs(std::move(pair));
     }
 
 protected: