+++ Moco/Moco/RegisterTypes_osimMoco.cpp
@@ -74,6 +74,7 @@ OSIMMOCO_API void RegisterTypes_osimMoco() {
         Object::registerType(MocoJointReactionGoal());
         Object::registerType(MocoOrientationTrackingGoal());
         Object::registerType(MocoTranslationTrackingGoal());
+        Object::registerType(MocoPeriodicityGoalPair());
         Object::registerType(MocoPeriodicityGoal());
         Object::registerType(MocoBounds());
         Object::registerType(MocoInitialBounds());