+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- 2019-07-18: Periodicity of the states and/or controls can be enforced with
+              MocoPeriodicityGoals that impose equality of the initial and final
+              state/control values.
 - 2019-07-15: MocoGoals can be enforced either as objective terms or as endpoint
               constraints. Applying a MocoGoal as an endpoint constraint is
               supported by MocoCasADiSolver (not MocoTropterSolver).